Implement a manifest.utils module
