+++ virtualizarr/manifests/utils.py
@@ -1,0 +1,118 @@
+from typing import TYPE_CHECKING, Any, Iterable, Union
+
+import numpy as np
+
+from virtualizarr.codecs import get_codecs
+
+if TYPE_CHECKING:
+    from zarr import Array  # type: ignore
+
+    from .array import ManifestArray
+
+
+def _check_same_dtypes(dtypes: list[np.dtype]) -> None:
+    """Check all the dtypes are the same"""
+
+    first_dtype, *other_dtypes = dtypes
+    for other_dtype in other_dtypes:
+        if other_dtype != first_dtype:
+            raise ValueError(
+                f"Cannot concatenate arrays with inconsistent dtypes: {other_dtype} vs {first_dtype}"
+            )
+
+
+def check_compatible_encodings(encoding1, encoding2):
+    for key, value in encoding1.items():
+        if key in encoding2:
+            if encoding2[key] != value:
+                raise ValueError(
+                    f"Cannot concatenate arrays with different values for encoding key {key}: {encoding2[key]} != {value}"
+                )
+
+
+def _check_same_codecs(codecs: list[Any]) -> None:
+    first_codec, *other_codecs = codecs
+    for codec in other_codecs:
+        if codec != first_codec:
+            raise NotImplementedError(
+                "The ManifestArray class cannot concatenate arrays which were stored using different codecs, "
+                f"But found codecs {first_codec} vs {codec} ."
+                "See https://github.com/zarr-developers/zarr-specs/issues/288"
+            )
+
+
+def _check_same_chunk_shapes(chunks_list: list[tuple[int, ...]]) -> None:
+    """Check all the chunk shapes are the same"""
+
+    first_chunks, *other_chunks_list = chunks_list
+    for other_chunks in other_chunks_list:
+        if other_chunks != first_chunks:
+            raise ValueError(
+                f"Cannot concatenate arrays with inconsistent chunk shapes: {other_chunks} vs {first_chunks} ."
+                "Requires ZEP003 (Variable-length Chunks)."
+            )
+
+
+def check_same_ndims(ndims: list[int]) -> None:
+    first_ndim, *other_ndims = ndims
+    for other_ndim in other_ndims:
+        if other_ndim != first_ndim:
+            raise ValueError(
+                f"Cannot concatenate arrays with differing number of dimensions: {first_ndim} vs {other_ndim}"
+            )
+
+
+def check_same_shapes(shapes: list[tuple[int, ...]]) -> None:
+    first_shape, *other_shapes = shapes
+    for other_shape in other_shapes:
+        if other_shape != first_shape:
+            raise ValueError(
+                f"Cannot concatenate arrays with differing shapes: {first_shape} vs {other_shape}"
+            )
+
+
+def _remove_element_at_position(t: tuple[int, ...], pos: int) -> tuple[int, ...]:
+    new_l = list(t)
+    new_l.pop(pos)
+    return tuple(new_l)
+
+
+def check_same_shapes_except_on_concat_axis(shapes: list[tuple[int, ...]], axis: int):
+    """Check that shapes are compatible for concatenation"""
+
+    shapes_without_concat_axis = [
+        _remove_element_at_position(shape, axis) for shape in shapes
+    ]
+
+    first_shape, *other_shapes = shapes_without_concat_axis
+    for other_shape in other_shapes:
+        if other_shape != first_shape:
+            raise ValueError(
+                f"Cannot concatenate arrays with shapes {[shape for shape in shapes]}"
+            )
+
+
+def check_combinable_zarr_arrays(
+    arrays: Iterable[Union["ManifestArray", "Array"]],
+) -> None:
+    """
+    The downside of the ManifestArray approach compared to the VirtualZarrArray concatenation proposal is that
+    the result must also be a single valid zarr array, implying that the inputs must have the same dtype, codec etc.
+    """
+    _check_same_dtypes([arr.dtype for arr in arrays])
+
+    # Can't combine different codecs in one manifest
+    # see https://github.com/zarr-developers/zarr-specs/issues/288
+    _check_same_codecs([get_codecs(arr) for arr in arrays])
+
+    # Would require variable-length chunks ZEP
+    _check_same_chunk_shapes([arr.chunks for arr in arrays])
+
+
+def check_compatible_arrays(
+    ma: "ManifestArray", existing_array: "Array", append_axis: int
+):
+    check_combinable_zarr_arrays([ma, existing_array])
+    check_same_ndims([ma.ndim, existing_array.ndim])
+    arr_shapes = [ma.shape, existing_array.shape]
+    check_same_shapes_except_on_concat_axis(arr_shapes, append_axis)
